Add basic authentication support and enable Flyway for database migration
