+++ api/app/src/main/resources/application-test.properties
@@ -2,6 +2,9 @@ spring.jpa.show-sql=false
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL82Dialect
 spring.jpa.properties.hibernate.format_sql=true
 packit.scheduling.enabled=false
+spring.jpa.hibernate.ddl-auto=validate
+spring.flyway.enabled=true
+spring.sql.init.mode=never
 #security
 spring.security.oauth2.client.registration.github.client-id=generateId
 spring.security.oauth2.client.registration.github.client-secret=generateSecret
